requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
langchain-core
langchain-text-splitters
chromadb
sentence-transformers
pypdf
streamlit
python-dotenv
torchvision